requirements.txt
crewai
google-generativeai>=0.4.0
python-dotenv
langchain-google-genai
crewai_tools